Comment and move vtable stuff into a module
